feat: move pages to (pages) to better organization
