+++ src/app/(pages)/favicon.ico
